Removed MLE package, LBFGS now in optimize
